requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
pydantic==2.5.2
tenacity==8.2.3
loguru==0.7.2 